vmnet: work around field-spanning write warning
The vmnet code uses struct VNet_EventHeader for an event header which is
followed by variable amount of payload data but does use a flexible array
member like most similar structures. When building with FORTIFY_SOURCE
against kernel 6.1-rc1, this results in runtime warnings like

  memcpy: detected field-spanning write (size 28) of single field "&t->event"

in VNetEvent_Send() and VNetUserListenerEventHandler(). Create a helper for
copying full event structure and for implement it using two separate copy
statements, one for fixed header and one for the variable payload. Another
approach seems to be the use of unsafe_memcpy() which might be more
appropriate as the final solution.